Switch from JttComponent to ServiceComponent
JttComponent was a static field in a class to ensure that all
dependencies it provides were singletons as long as application process
exists. ServiceComponent is only a singleton within JttService class and
it is destroyed as soon as service is stopped (which should never
happen).

The only dependency that actually has to be a singleton is the Ticker.
SolarEventCalculator is a singleton for now since it must be the same
object when it is used for calculations and when it is provided the
local coordinates. This will be fixed when LocationProvider will be
introduced.