JttService now uses startForeground
This means the service can be stopped if notification is not
shown. The problem is that it handles broadcast listeners for
screen_on and screen_off which means without notification the
application can become unusable.

The solution for now is to make the service always running by having
the notification always shown.